Some refactoring per review comments.
